Add render::common module
This stores `Resolve` and `Render` impls for top-level elements like
`Variable`, `TokenTree` and `Text` and elements used by both tags and
filters, like `Argument` and `TagElement`.
